Formatting date with handlebars.moment

+++ models/stocks.js
@@ -23,7 +23,7 @@ module.exports = (sequelize, DataTypes) => {
       allowNull: false,
     },
     stock_price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.FLOAT(6, 2),
       allowNull: false,
     },
     stock_quantity: {
